+++ services/billing-core/src/main/java/br/com/arinelli/pay/billing/people/NaturalPersonRepository.java
@@ -1,0 +1,8 @@
+package br.com.arinelli.pay.billing.people;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+interface NaturalPersonRepository extends JpaRepository<NaturalPerson, Long> {
+
+    boolean existsByCpf(String cpf);
+}